MicroService/notification-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-jose
psycopg2-binary
python-dotenv
python-multipart
